+++ src/Network/Network.Test/ScenarioTests/ApplicationGatewayTests.ps1
@@ -1252,6 +1252,169 @@ function Test-ApplicationGatewayCRUDRewriteRuleSetWithUrlConfiguration
 .SYNOPSIS
 Application gateway Basic SKU tests
 #>
+function Test-ApplicationGatewayGlobalConfig
+{
+	param
+	(
+		$basedir = "./"
+	)
+
+	# Setup
+	$location = Get-ProviderLocation "Microsoft.Network/applicationGateways" "East US"
+
+	$rgname = Get-ResourceGroupName
+	$appgwName = Get-ResourceName
+	$identityName = Get-ResourceName
+	$vnetName = Get-ResourceName
+	$gwSubnetName = Get-ResourceName
+	$publicIpName = Get-ResourceName
+	$gipconfigname = Get-ResourceName
+
+	$frontendPort01Name = Get-ResourceName
+	$fipconfigName = Get-ResourceName
+	$listener01Name = Get-ResourceName
+
+	$poolName = Get-ResourceName
+	$trustedRootCertName = Get-ResourceName
+	$poolSetting01Name = Get-ResourceName
+
+	$rule01Name = Get-ResourceName
+
+	$probeHttpName = Get-ResourceName
+
+	try
+	{
+		# Create the resource group
+		$resourceGroup = New-AzResourceGroup -Name $rgname -Location $location -Tags @{ testtag = "APPGw tag"}
+		# Create the Virtual Network
+		$gwSubnet = New-AzVirtualNetworkSubnetConfig -Name $gwSubnetName -AddressPrefix 10.0.0.0/24
+		$vnet = New-AzVirtualNetwork -Name $vnetName -ResourceGroupName $rgname -Location $location -AddressPrefix 10.0.0.0/16 -Subnet $gwSubnet
+		$vnet = Get-AzVirtualNetwork -Name $vnetName -ResourceGroupName $rgname
+		$gwSubnet = Get-AzVirtualNetworkSubnetConfig -Name $gwSubnetName -VirtualNetwork $vnet
+
+		# Create public ip
+		$publicip = New-AzPublicIpAddress -ResourceGroupName $rgname -name $publicIpName -location $location -AllocationMethod Static -sku Standard
+
+		# Create ip configuration
+		$gipconfig = New-AzApplicationGatewayIPConfiguration -Name $gipconfigname -Subnet $gwSubnet
+
+		$fipconfig = New-AzApplicationGatewayFrontendIPConfig -Name $fipconfigName -PublicIPAddress $publicip
+		$fp01 = New-AzApplicationGatewayFrontendPort -Name $frontendPort01Name  -Port 80
+		$listener01 = New-AzApplicationGatewayHttpListener -Name $listener01Name -Protocol Http -FrontendIPConfiguration $fipconfig -FrontendPort $fp01
+
+		# backend part
+		# trusted root cert part
+		$certFilePath = $basedir + "/ScenarioTests/Data/ApplicationGatewayAuthCert.cer"
+		$trustedRoot01 = New-AzApplicationGatewayTrustedRootCertificate -Name $trustedRootCertName -CertificateFile $certFilePath
+		$pool = New-AzApplicationGatewayBackendAddressPool -Name $poolName -BackendIPAddresses www.microsoft.com, www.bing.com
+		$probeHttp = New-AzApplicationGatewayProbeConfig -Name $probeHttpName -Protocol Https -HostName "probe.com" -Path "/path/path.htm" -Interval 89 -Timeout 88 -UnhealthyThreshold 8
+		$poolSetting01 = New-AzApplicationGatewayBackendHttpSettings -Name $poolSetting01Name -Port 443 -Protocol Https -Probe $probeHttp -CookieBasedAffinity Enabled -PickHostNameFromBackendAddress -TrustedRootCertificate $trustedRoot01
+
+		#rule
+		$rule01 = New-AzApplicationGatewayRequestRoutingRule -Name $rule01Name -RuleType basic -Priority 100 -BackendHttpSettings $poolSetting01 -HttpListener $listener01 -BackendAddressPool $pool
+
+		# sku
+		$sku = New-AzApplicationGatewaySku -Name Standard_v2 -Tier Standard_v2 -Capacity 2
+
+		# Create Application Gateway
+		$appgw = New-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Location $location -Probes $probeHttp -BackendAddressPools $pool -BackendHttpSettingsCollection $poolSetting01 -FrontendIpConfigurations $fipconfig -GatewayIpConfigurations $gipconfig -FrontendPorts $fp01 -HttpListeners $listener01 -RequestRoutingRules $rule01 -Sku $sku -TrustedRootCertificate $trustedRoot01
+
+		# Get Application Gateway
+		$getgw = Get-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname
+
+		# Operational State
+		Assert-AreEqual "Running" $getgw.OperationalState
+
+		# check trusted root
+		$trustedRoot02 = Get-AzApplicationGatewayTrustedRootCertificate -ApplicationGateway $getgw -Name $trustedRootCertName
+		Assert-NotNull $trustedRoot02
+		Assert-AreEqual $getgw.BackendHttpSettingsCollection[0].TrustedRootCertificates.Count 1
+
+		# check sku
+		$sku01 = Get-AzApplicationGatewaySku -ApplicationGateway $getgw
+		Assert-NotNull $sku01
+		Assert-AreEqual $sku01.Capacity 2
+		Assert-AreEqual $sku01.Name Standard_v2
+		Assert-AreEqual $sku01.Tier Standard_v2
+
+		# Assert by default Global Configuration is null
+		$globalConfiguration = Get-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw
+		Assert-Null $globalConfiguration
+		
+		# Set only EnableRequestBuffering to false and assert that EnableResponseBuffering is still null
+		Set-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw -EnableRequestBuffering $false
+		$getgw = Set-AzApplicationGateway -ApplicationGateway $getgw
+
+		$globalConfiguration = Get-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw
+		Assert-NotNull $globalConfiguration
+		Assert-AreEqual $false $globalConfiguration.EnableRequestBuffering
+		Assert-Null $globalConfiguration.EnableResponseBuffering
+
+		# Now set EnableRequestBuffering and EnableResponseBuffering to true
+		Set-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw -EnableRequestBuffering $true -EnableResponseBuffering $true
+		$getgw = Set-AzApplicationGateway -ApplicationGateway $getgw
+		$getgw = Get-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname
+		$globalConfiguration = Get-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw
+		Assert-NotNull $globalConfiguration
+		Assert-AreEqual $true $globalConfiguration.EnableRequestBuffering
+		Assert-AreEqual $true $globalConfiguration.EnableResponseBuffering
+
+		# Remove the Globabl Configuration and assert that the property is null
+		Remove-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw
+		Set-AzApplicationGateway -ApplicationGateway $getgw
+		$getgw = Get-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname
+		$globalConfiguration = Get-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw
+		Assert-Null $globalConfiguration
+
+		# Stop Application Gateway
+		$getgwStopped = Stop-AzApplicationGateway -ApplicationGateway $getgw
+
+		Assert-AreEqual "Stopped" $getgwStopped.OperationalState
+
+		# Delete Application Gateway
+		Remove-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Force
+
+		# Create Application Gateway with Global Configuration and assert that the values are set properly
+		$globalConfig = New-AzApplicationGatewayGlobalConfiguration -EnableRequestBuffering $false -EnableResponseBuffering $true
+		$getgw2 = New-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Location $location -Probes $probeHttp -BackendAddressPools $pool -BackendHttpSettingsCollection $poolSetting01 -FrontendIpConfigurations $fipconfig -GatewayIpConfigurations $gipconfig -FrontendPorts $fp01 -HttpListeners $listener01 -RequestRoutingRules $rule01 -Sku $sku -TrustedRootCertificate $trustedRoot01 -GlobalConfiguration $globalConfig
+		
+		$getgw3 = Get-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname
+		Assert-AreEqual "Running" $getgw3.OperationalState
+
+		$getGlobalConfig = Get-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw3
+		Assert-NotNull $getGlobalConfig
+		Assert-AreEqual $false $getGlobalConfig.EnableRequestBuffering
+		Assert-AreEqual $true $getGlobalConfig.EnableResponseBuffering
+
+		# Assert that Global Configuration remains after other powershell commands are executed (ensure bug is fixed)
+		$getgw3 = Set-AzApplicationGatewayBackendAddressPool -ApplicationGateway $getgw3 -Name $poolName -BackendIPAddresses www.microsoft.com, www.bing.com, www.amazon.com
+		Set-AzApplicationGateway -ApplicationGateway $getgw3
+		$getgw3 = Get-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname
+
+		$getGlobalConfig = Get-AzApplicationGatewayGlobalConfiguration -ApplicationGateway $getgw3
+		Assert-NotNull $getGlobalConfig
+		Assert-AreEqual $false $getGlobalConfig.EnableRequestBuffering
+		Assert-AreEqual $true $getGlobalConfig.EnableResponseBuffering
+
+		# Stop Application Gateway
+		$getgw2Stopped = Stop-AzApplicationGateway -ApplicationGateway $getgw3
+
+		Assert-AreEqual "Stopped" $getgw2Stopped.OperationalState
+
+		# Delete Application Gateway
+		Remove-AzApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Force
+	}
+	finally
+	{
+		# Cleanup
+		Clean-ResourceGroup $rgname
+	}
+}
+
+<#
+.SYNOPSIS
+Application gateway Basic SKU tests
+#>
 function Test-ApplicationGatewayBasicSkuCRUD
 {
 	param