+++ src/Network/Network/ChangeLog.md
@@ -25,6 +25,7 @@
 * Added cmdlet `New-AzSaaSNetworkVirtualAppliance` for creating a NetworkVirtualAppliance of SaaS type.
 * Added control knobs to virtual network gateways and ExpressRoute gateways as well to cmdlets operating on those.
 * Added cmdlet 'Remove-AzApplicationGatewayFirewallCustomRule' to support removing custom rule in Firewall Policy.
+* Added support for Application Gateway GlobalConfiguration
 
 ## Version 6.2.0
 * Added support for new Application Gateway SKU type, Basic SKU